Add a widget to configure the network
To get data, the model uses NetworkManagerQt. Once a second, a list of
all connections is requested, and changes are made to the model.

Implemented methods for enabling and disabling Wifi and Networking,
adding and removing temporary Wifi connections with open or closed
access points with a pre-shared key.

Since connections are temporary, they are lost after a reboot.