Update SelectImageButton component
Show an image instead of a button to open a dialog with this image.
When hovering the mouse on the image, there will be buttons for
selecting a file and deleting.

Also moved SelectImageButton.qml to the kcm directory, so that there are
no problems with translated messages, and since this component is used
specifically for the configuration.