Show packaged images in KCM and greeter
The org.kde.plasma.wallpaper.image package has a MediaProxy component
that makes it easy to determine the image source by url or package
folder. We use this to display a preview on the SelectImageButton.

But since we don't want to use this in the greeter, part of the code was
copied, to determine the preferred image filename from the url, when
saved in configuration module.

Also, instead of the KeepAspectRatio setting, the FillMode setting has
been made, which allows you to choose from several modes for displaying
the desktop background.