Copy all needed images to greeter user folder
In the customization utility, each theme should be able to save some
images and later use them when logging in. Since the greeter is running
as a separate limited user _ldm, it cannot see the user's files.

In order for the greeter to see the selected images, applying the
settings will copy those files to the greeter's home directory, since
privileges are needed to apply the settings.

Add the path to the home directory of the greeter user as a compilation
option. Block the ability to build without specifying the
GREETER_IMAGES_DIR variable, since this is now required information, not
implying a default value.