Disable rootimageapp
It is a special separate app to prevent a black flicker when logging in
via LightDM-KDE before KSplash starts. Apparently, it has lost its
relevance, and now leads to long freezes when logging in. Removing this
application does not cause additional flickering, but significantly
speeds up the login. The application will be left, but will only be
activated if the enable-root-image-app=true option is present.