Improve config save/load system - part 2
Each setting is now associated with an object of type ConfigValue. There
you can specify the path in the config file, type, default value. And
also from which widget you need to "listen" to the value.

Each time the listened widget is changed, the entire config is checked
for compliance with the original version of the config. And depending on
this, the so-called "NeedsSave" is activated or deactivated. Therefore,
if the user made a chain of changes that led to the initial state, we
will know for sure that nothing needs to be saved.