Add a service to prolong wifi connection
If the user session reuses the X server, the greeter session ends before
the user session starts.

When greeter's session ends, the configured wifi connection is disconnected.

Therefore, it is not possible to log in as a domain user if his data is
not yet in the sss cache.

A separate "wifi-keeper" service was added, through which a connection
is created, and it lives a little longer than the greeter, but this is
enough for the domain user to log in successfully.