Support "hide-users" hint
If the hide-users option is enabled, we immediately begin authentication
and the user sees a name prompt. And hide the "cancel" button in this
prompt, because there is nothing to cancel.

Also, don't take into account the approximate size of the button if it's
not visible. Now this matters because sometimes the "Login as another
user" button will not be visible.