Provide some time to read messages from PAM
Sometimes the PAM module sends some messages immediately before
successful authentication, and the user barely has time to see them.
Added a simple mechanism that allows to slightly delay the start of the
session so that the user can read all messages.

Looked in GDM, the time it takes to view a message is calculated as 48
milliseconds per character. Implemented similar here.

Also, added that if an identical message is already in the list, extra
time is not allocated. This has proven useful for fingerprint
authentication, for example; it spams you with repeated messages if it
fails to read your fingerprint. Which results in an excessively long
timeout after authentication.

Fixes: https://bugzilla.altlinux.org/51493