Connect to a VPN that requires entering secrets
Add a class inherited from NetworkManager::SecretAgent to create a dbus
interface for receiving the request and passing secrets to
NetworkManager.

Add a dialog layout for requesting secrets from the user, with a
variable number of input fields, according to the number of required
secrets.

Add an error message indicating the cause of the error, in particular to
display an error when connecting to a VPN.

OpenVPN is primarily supported