Remove the main window that is not visible
A long time ago, all the graphics were displayed in one window stretched
on all physical screens. It caused problems, and it was redone so that
ts own window was shown on each screen. But the main window remained and
was hidden by working windows. It does not needed at all, and it can be
seen when animation at the start, because it is filled with white color.
Therefore, it will be removed, only visible windows will remain, one for
each physical screen.