Add KWin backend for KeyboardModel
KWin provides "org.kde.keyboard" DBus service, through which, for
example, you can get a list of layouts and set the current. There is a
QML component in org.kde.plasma.workspace.components,
KeyboardLayoutSwitcher, which internally uses the specified service.

It is not very convenient to use here, since it will not work under X11,
and we still need this. On the other hand, our KeyboardModel supports
replaceable backens that abstract interaction with a display server
regarding the keyboard.

Therefore, another backend (KWinKeyboardBackend) was written, most of
which is borrowed from the implementation of the aforementioned QML
component.